// src/types/scopes.d.ts
export type Scope =
	"analytics:read:extensions" |
	"analytics:read:games" |
	"bits:read" |
	"channel:edit:commercial" |
	"channel:read:hype_train" |
	"channel:read:subscriptions" |
	"channel:read:stream_key" |
	"clips:edit" |
	"user:edit" |
	"user:edit:broadcast" |
	"user:edit:follows" |
	"user:read:broadcast" |
	"user:read:email" |
	"moderation:read";